Cleanup and split provider code into its own package
